move properties widget collapse to bottom
