+++ src/sql/workbench/contrib/dashboard/browser/contents/dashboardWidgetWrapper.component.html
@@ -18,4 +18,5 @@
 		<ng-template component-host>
 		</ng-template>
 	</ng-template>
+	<span #bottomActionbar style="flex: 0 0 auto; align-self: center;"></span>
 </div>
