+++ src/sql/workbench/contrib/dashboard/browser/pages/databaseDashboardPage.component.ts
@@ -14,7 +14,6 @@ import { CommonServiceInterface } from 'sql/workbench/services/bootstrap/browser
 import { IAngularEventingService } from 'sql/platform/angularEventing/browser/angularEventingService';
 
 import * as colors from 'vs/platform/theme/common/colorRegistry';
-import * as nls from 'vs/nls';
 import { INotificationService } from 'vs/platform/notification/common/notification';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { ILogService } from 'vs/platform/log/common/log';
@@ -27,7 +26,6 @@ import { IWorkbenchThemeService } from 'vs/workbench/services/themes/common/work
 
 export class DatabaseDashboardPage extends DashboardPage implements OnInit {
 	protected propertiesWidget: WidgetConfig = {
-		name: nls.localize('databasePageName', "DATABASE DASHBOARD"),
 		widget: {
 			'properties-widget': undefined
 		},
